Refactor switch creation, for using switch invariants
To make things much simpler than what we had (with lots of code to
choose the compilers when creating switches on existing projects), the
switch creation is now done in a single pass (compiler + local
packages), which means the single call to the solver determines the
compiler: this should be much more reliable.

Then the "switch invariant" is set according to the packages with the
"compiler" flag that got installed.

Note that this should also be much more friendly when creating a switch
in a directory that contains compiler definitions (this is detected, and
the "default compiler" selection isn't used).