Overhaul OpamStd.Sys Cygwin functions
Make the ~cygbin parameter to the various executable identification
functions optional. The parameter is renamed to search_in_first with a
slightly revised semantics - if cygcheck.exe is not found in the
directory, then PATH is still searched.

This has a key benefit early on in opam init, as it allows cygbin to be
set for an internal Cygwin installation which has not yet happened, but
still permits curl (et al) to be used from PATH (e.g. when running opam
init from an MSYS2 shell).

Where before ~cygbin implied that cygcheck _must_ be used from the
directory, the new parameter effectively prepends an additional
directory to PATH.