Detect both Cygwin and MSYS2 os-distribution
Previously, MSYS2 required os-distribution to be overridden in
global-variables. Now, as with Cygwin, it is inferred in the same way
from cygcheck.

In this commit, opam init still sets os-distribution, however, if it is
manually removed from the root config file, previously `opam var
os-distribution` would return `win32` but now returns `msys2`.

Additionally, sys_pkg_manager_cmd is handled in the same way for MSYS2
as for Cygwin, allowing MSYS2 to be automatically added in the same way
as Cygwin.