Propagate parse errors correctly
opam-file-format has not required a change since opam 2.1, which causes
it to report parse errors when opam-version is > 2.1 using a signal.
opam was using the format version instead of the client version to
determine if these errors are actually parse errors, and was therefore
reporting an unsupported format instead of a parse error.